fixing or supressing various warnings
* adding explicit casts or types  to some numeric operations
  - the warning caught a potential overflow bug, although one we were very unlikely to ever hit
* supressing 'this-escape' warnings in GATKTool because they are an inevitable feature of the design
* making several methods in GATKPath final to avoid the same warning

+++ src/main/java/org/broadinstitute/hellbender/tools/AnalyzeSaturationMutagenesis.java
@@ -698,7 +698,7 @@ public final class AnalyzeSaturationMutagenesis extends GATKTool {
             final SimpleInterval wholeTig = new SimpleInterval(tig0.getSequenceName(), 1, refSeqLen);
             refSeq = Arrays.copyOf(refSource.queryAndPrefetch(wholeTig).getBases(), refSeqLen);
             for ( int idx = 0; idx < refSeqLen; ++idx ) {
-                switch ( refSeq[idx] &= UPPERCASE_MASK ) { // make into upper case
+                switch ( refSeq[idx] &= (byte) UPPERCASE_MASK) { // make into upper case
                     case 'A':
                     case 'C':
                     case 'G':
